fixed warning
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cado-nfs/trunk@671 3eaf19af-ecc0-40f6-b43f-672aa0c71c71

+++ linalg/bwc/balancing_rollbuf.c
@@ -201,7 +201,7 @@ int rollbuf_get(rollbuf_ptr r, char * p, size_t s)
 #ifndef NDEBUG
         ptrdiff_t head = r->avail_to_read - tail;
         assert(head > 0);
-        assert((size_t) (s - tail) <= head);
+        assert((size_t) (s - tail) <= (size_t) head);
         // s = tail + MIN((size_t) head, s - tail);
 #endif
         assert(s >= tail);
